support for float seconds

+++ src/ArrivalDepartureTime.ts
@@ -7,7 +7,7 @@ export class ArrivalDepartureTime {
 
     private parse(offset: number): number {
         const split = this.value.split(/([-+])/);
-        return parseInt(split[offset]) * (split[offset-1] == '-' ? -1 : 1)
+        return parseFloat(split[offset]) * (split[offset-1] == '-' ? -1 : 1)
     }
 
     get departure(): number {
